37-cadence-pme-adaptive-tutor/requirements.txt
streamlit>=1.30
plotly>=5.18
pandas>=2.0
openai>=1.40
python-dotenv>=1.0
python-docx>=1.1
openpyxl>=3.1
imageio-ffmpeg>=0.5.1
